# config/seng365-2018_auction_0.0.12_swagger.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "info": {
    "description": "",
    "version": "0.0.12",
    "title": "Auction API for SENG365 2018 Assignments"
  },
...
  "paths": {
...
    "/auctions": {
      "get": {
        "tags": [
          "auctions"
        ],
        "summary": "View auctions, sorted from most recent to least recent.",
        "parameters": [
          {
            "in": "query",
            "name": "startIndex",
            "description": "Number of items to skip before returning result",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "in": "query",
            "name": "count",
            "description": "Number of items to return",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "in": "query",
            "name": "q",
            "description": "Search string for auction titles",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "category-id",
            "description": "ID for the category.",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "in": "query",
            "name": "seller",
            "description": "Only return auctions with this id as a seller",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "in": "query",
            "name": "bidder",
            "description": "Only return auctions which have a bid by this id",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "in": "query",
            "name": "winner",
            "description": "Only return auctions which were won by this id",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "in": "query",
            "name": "status",
            "description": "Return auctions with a certain status. Can be 'all', 'active', 'expired', 'upcoming' or 'won'",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/auctionsOverview"
                }
              }
            }
          },
          "400": {
            "description": "Bad request."
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "auctions"
        ],
        "summary": "Create auction",
        "security": [
          {
            "User": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "title",
                  "categoryId",
                  "description",
                  "startDateTime",
                  "endDateTime",
                  "reservePrice",
                  "startingBid"
                ],
                "allOf": [
                  {
                    "$ref": "#/components/schemas/categoryId"
                  },
                  {
                    "$ref": "#/components/schemas/title"
                  },
                  {
                    "$ref": "#/components/schemas/description"
                  },
                  {
                    "$ref": "#/components/schemas/startDateTime"
                  },
                  {
                    "$ref": "#/components/schemas/endDateTime"
                  },
                  {
                    "$ref": "#/components/schemas/reservePrice"
                  },
                  {
                    "$ref": "#/components/schemas/startingBid"
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/id"
                }
              }
            }
          },
          "400": {
            "description": "Bad request."
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "auctionsOverview": {
        "type": "array",
        "items": {
          "type": "object",
          "required": [
            "id",
            "title"
          ],
          "description": "summary data for an Auction",
          "allOf": [
            {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer"
                },
                "categoryTitle": {
                  "type": "string"
                }
              }
            },
            {
              "$ref": "#/components/schemas/categoryId"
            },
            {
              "$ref": "#/components/schemas/title"
            },
            {
              "$ref": "#/components/schemas/reservePrice"
            },
            {
              "$ref": "#/components/schemas/startDateTime"
            },
            {
              "$ref": "#/components/schemas/endDateTime"
            },
            {
              "type": "object",
              "properties": {
                "currentBid": {
                  "type": "number",
                  "minimum": 0
                }
              }
            }
          ]
        }
      },
...
      "id": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "categoryId": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "title": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          }
        }
      },
      "reservePrice": {
        "type": "object",
        "properties": {
          "reservePrice": {
            "type": "number"
          }
        }
      },
      "startDateTime": {
        "type": "object",
        "properties": {
          "startDateTime": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "endDateTime": {
        "type": "object",
        "properties": {
          "endDateTime": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "description": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          }
        }
      },
      "startingBid": {
        "type": "object",
        "properties": {
          "startingBid": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
    },
    "securitySchemes": {
      "User": {
        "type": "apiKey",
        "name": "X-Authorization",
        "in": "header"
      }
    }
  }
}